refactored group management functions
